Extracted decorator out to a separate file, added decorator experimental flag to gulp task
